bool state to separate concerns [12:19am]
if not, 
the select icon proceeds in a way that the user profileIcon
will show before they confirm. 

This state gets in the way of that state being the only decider
because that state (in the way i was using it) is needed 
to press along in the steps with the promise and to save the state

to persist through changing components.
same state localStorage is based on for localStorage.