I M && U R [1:47pm]
<settings/> screen width gets low in px ---> text doesn't fit.


"imperial" | "metric" 
the state that determines which units will be used. 

"update" | "reset"
btn: decide whether user settings state/ DB will stay or reset.
Update btn is a post route that updates the users settings.

decided to change:
Imperial and Metric to " I " && " M " 
update and reset to " U " && " R " 

also seems lucky of an idea to play with because:
the settings are personal. 
to have the shorthand for some of these personal settings be:
I M ( i am ) UR ( you are ) 

falls in line with the rest of the app's whimsicality in my eyes.